+++ back/src/main/java/com/example/back/repositorios/CitaRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.back.Cita;
 
 
-public interface CitaRepositorio extends JpaRepository<Cita, Integer>{
+public interface CitaRepositorio extends JpaRepository<Cita, Long>{
     
 }
