+++ back/src/main/java/com/example/back/repositorios/ImagenesRepositorio.java
@@ -3,6 +3,6 @@ package com.example.back.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.back.Imagenes;
 
-public interface ImagenesRepositorio extends JpaRepository<Imagenes, Integer>{
+public interface ImagenesRepositorio extends JpaRepository<Imagenes, Long>{
     
 }
