+++ back/src/main/java/com/example/back/repositorios/ServicioRepositorio.java
@@ -3,6 +3,6 @@ package com.example.back.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.back.Servicio;
 
-public interface ServicioRepositorio extends JpaRepository<Servicio, Integer>{
+public interface ServicioRepositorio extends JpaRepository<Servicio, Long>{
 
 }
