+++ back/src/main/java/com/example/back/servicios/ClienteService.java
@@ -6,12 +6,12 @@ import com.example.back.Cliente;
 
 public interface ClienteService{
     public List<Cliente> getAllClientes();
-    public Cliente getClienteById(Integer id);
+    public Cliente getClienteById(Long id);
     public Cliente guardarCliente(Cliente cliente);
-    public Cliente actualizarCliente(Integer id, Cliente cliente);
-    public void eliminarCliente(Integer id);    
+    public Cliente actualizarCliente(Long id, Cliente cliente);
+    public void eliminarCliente(Long id);    
     public Cliente login(String email, String contrasena);
     public Cliente getClienteByEmail(String email);//para recuperar la contraseña
-    public Cliente actualizarContrasena(Integer id, String nuevaContrasena);//para actualizarla
+    public Cliente actualizarContrasena(Long id, String nuevaContrasena);//para actualizarla
 
 }
